requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.11.2

# AWS CDK dependencies
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# Lambda function dependencies
boto3>=1.26.0
s3fs>=2023.12.0
dask>=2023.12.0
xarray>=2023.12.0
zarr>=2.16.0
netCDF4>=1.6.5

# Development dependencies
localstack>=2.3.0
pytest>=7.0.0
pytest-mock>=3.10.0
moto>=4.0.0